Implement crude automatic installation procedure
The makefile is pretty gross and I'm 100% confident that
you can use m4 or some other templating tool for what I
did with sed, but this does seem to work.

An important note is that this now runs as a different
group (uinput), which means that the raw device files
are no longer readable by the user. This is good, as it
prevents a program running as the user from recording all
keys on the raw device file.